pyproject.toml
[project]
name = "agent-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=7.5.1",
    "groq>=0.18.0",
    "phidata>=2.7.10",
    "yfinance>=0.2.54",
]
